fix a compile warning.
* vm/vm_page.c(vm_page_setup): %lu -> %zu

vm/vm_page.c: In function 'vm_page_setup':
vm/vm_page.c:1425:41: warning: format '%lu' expects argument of type 'long unsigned int', but argument 2 has type 'size_t' {aka 'unsigned int'} [-Wformat=]
 1425 |     printf("vm_page: page table size: %lu entries (%luk)\n", nr_pages,
      |                                       ~~^                    ~~~~~~~~
      |                                         |                    |
      |                                         long unsigned int    size_t {aka unsigned int}
      |                                       %u
vm/vm_page.c:1425:54: warning: format '%lu' expects argument of type 'long unsigned int', but argument 3 has type 'size_t' {aka 'unsigned int'} [-Wformat=]
 1425 |     printf("vm_page: page table size: %lu entries (%luk)\n", nr_pages,
      |                                                    ~~^
      |                                                      |
      |                                                      long unsigned int
      |                                                    %u
 1426 |            table_size >> 10);
      |            ~~~~~~~~~~~~~~~~
      |                       |
      |                       size_t {aka unsigned int}
Message-ID: <20241020190744.2522-1-jbranso@dismail.de>

+++ vm/vm_page.c
@@ -1422,8 +1422,9 @@ vm_page_setup(void)
         nr_pages += vm_page_atop(vm_page_boot_seg_size(&vm_page_boot_segs[i]));
 
     table_size = vm_page_round(nr_pages * sizeof(struct vm_page));
-    printf("vm_page: page table size: %lu entries (%luk)\n", nr_pages,
-           table_size >> 10);
+    printf("vm_page: page table size: %lu entries (%luk)\n",
+	   (unsigned long) nr_pages,
+	   (unsigned long) (table_size >> 10));
     table = (struct vm_page *)pmap_steal_memory(table_size);
     va = (unsigned long)table;
